+++ push_swap/libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: svikornv <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: svikornv <svikornv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 14:31:54 by svikornv          #+#    #+#             */
-/*   Updated: 2023/06/25 15:56:02 by svikornv         ###   ########.fr       */
+/*   Updated: 2023/08/04 11:36:56 by svikornv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ typedef struct s_list
 The first part of project.
 */
 int		ft_isalpha(int c);
-int		ft_isdigit(int c);
+int		ft_isdigit(long c);
 int		ft_isalnum(int c);
 int		ft_isascii(int c);
 int		ft_isprint(int c);
@@ -50,7 +50,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n);
 void	*ft_memchr(const void *s, int c, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-int		ft_atoi(const char *s);
+long		ft_atoi(const char *s);
 void	*ft_calloc(size_t count, size_t size);
 char	*ft_strdup(const char *s);
